pyproject.toml
[tool.poetry]
name = "futbol-bot"
version = "0.1.0"
description = "Bot Telegram 1v1 avec PostgreSQL"
authors = ["Horeb"]
license = "MIT"
packages = [
  { include = "bot.py" }
]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "20.8"
psycopg2-binary = "^2.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
